แก้activity

+++ ProjectManagesy_BE/app/Repositories/ActivityRepository.php
@@ -14,40 +14,32 @@ class ActivityRepository implements ActivityRepositoryInterface
 {
     public function createActivity($data)
     {
-
         $activity = new ActivityStudent;
-        // $activity->activity_student_name = $data['activity_student_name'];
-        // $activity->activity_student_year = $data['activity_student_year'];
-        // $activity->activity_student_major = $data['activity_student_major'];
-        // $activity->activity_student_file_name = $data['activity_student_file_name'];
-        // $activity->save();
+        $activity->activity_student_name = $data['activity_student_name'];
+        $activity->activity_student_year = $data['activity_student_year'];
+        $activity->activity_student_major = $data['activity_student_major'];
+        $activity->activity_student_file_name = $data['activity_student_file_name'];
+        $activity->save();
 
-        foreach ($data['activity_student_file'] as  $value) {
-            if (
-                $value['data_first_name'] && $value['data_surname']
-                && $value['data_degree'] && $value['data_school_name']
-                && $value['data_email'] && $value['data_tel']
-            ) {
-
-                // $activity = new ActivityStudent;
-                $activity->activity_student_name = $data['activity_student_name'];
-                $activity->activity_student_year = $data['activity_student_year'];
-                $activity->activity_student_major = $data['activity_student_major'];
-                $activity->activity_student_file_name = $data['activity_student_file_name'];
-                $activity->save();
-
-
-                $activity_student_file = new ActivityStudentFile();
-                $activity_student_file->data_first_name = $value['data_first_name'];
-                $activity_student_file->data_surname = $value['data_surname'];
-                $activity_student_file->data_degree = $value['data_degree'];
-                $activity_student_file->data_school_name = $value['data_school_name'];
-                $activity_student_file->data_email = $value['data_email'];
-                $activity_student_file->data_tel = $value['data_tel'];
-                $activity_student_file->activity_student_id = $activity->id;
-                $activity_student_file->save();
-            } else {
-                ActivityStudent::where('activity_student_id', $activity->id)->delete();
+        if ($data['activity_student_file']) {
+            foreach ($data['activity_student_file'] as  $value) {
+                if (
+                    $value['data_first_name'] && $value['data_surname']
+                    && $value['data_degree'] && $value['data_school_name']
+                    && $value['data_email'] && $value['data_tel']
+                ) {
+                    $activity_student_file = new ActivityStudentFile();
+                    $activity_student_file->data_first_name = $value['data_first_name'];
+                    $activity_student_file->data_surname = $value['data_surname'];
+                    $activity_student_file->data_degree = $value['data_degree'];
+                    $activity_student_file->data_school_name = $value['data_school_name'];
+                    $activity_student_file->data_email = $value['data_email'];
+                    $activity_student_file->data_tel = $value['data_tel'];
+                    $activity_student_file->activity_student_id = $activity->id;
+                    $activity_student_file->save();
+                } else {
+                    ActivityStudent::where('activity_student_id', $activity->id)->delete();
+                }
             }
         }
     }
@@ -67,13 +59,13 @@ class ActivityRepository implements ActivityRepositoryInterface
                 ->update([
                     'activity_student.activity_student_file_name' => $data['activity_student_file_name']
                 ]);
-            ActivityStudentFile::where('activity_student_id', $data['activity_student_id'])->delete();
-            foreach ($data['activity_student_file'] as  $value) {
+            foreach ($data['activity_student_file'] as $value) {
                 if (
                     $value['data_first_name'] && $value['data_surname']
                     && $value['data_degree'] && $value['data_school_name']
                     && $value['data_email'] && $value['data_tel']
                 ) {
+                    ActivityStudentFile::where('activity_student_id', $data['activity_student_id'])->delete();
                     $activity_student_file = new ActivityStudentFile();
                     $activity_student_file->data_first_name = $value['data_first_name'];
                     $activity_student_file->data_surname = $value['data_surname'];
@@ -84,11 +76,14 @@ class ActivityRepository implements ActivityRepositoryInterface
                     $activity_student_file->activity_student_id = $data['activity_student_id'];
                     $activity_student_file->save();
                 } else {
-                    ActivityStudent::where('activity_student_id', $data['activity_student_id'])->update([
-                        'activity_student.activity_student_name' => $activity_old->activity_student_name,
-                        'activity_student.activity_student_year' => $activity_old->activity_student_year,
-                        'activity_student.activity_student_major' => $activity_old->activity_student_major
-                    ]);
+                    ActivityStudent::where('activity_student_id', $data['activity_student_id'])
+                        ->update([
+                            'activity_student.activity_student_name' => $activity_old->activity_student_name,
+                            'activity_student.activity_student_year' => $activity_old->activity_student_year,
+                            'activity_student.activity_student_major' => $activity_old->activity_student_major,
+                            'activity_student.activity_student_file_name' => $activity_old->activity_student_file_name
+                        ]);
+                    return 'fail';
                 }
             }
         }
@@ -100,7 +95,7 @@ class ActivityRepository implements ActivityRepositoryInterface
 
         foreach ($activity as $value) {
             $activity_file = ActivityStudentFile::where('activity_student_id', $value['activity_student_id'])->get();
-            $value['acitivty_file'] = $activity_file;
+            $value['activity_file'] = $activity_file;
         }
 
         return $activity;
